Move InputState next to UIScreen
From the user perspective the UIScreen is the only class using this
enum.
