+++ examples/entry_counter/entry_counter.py
@@ -1,8 +1,7 @@
 #!/bin/python3
 
 from simpleline import App
-from simpleline.render import InputState
-from simpleline.render.screen import UIScreen
+from simpleline.render.screen import UIScreen, InputState
 from simpleline.render.widgets import TextWidget, CenterWidget
