+++ examples/ignore_continue/ignore_continue.py
@@ -1,9 +1,8 @@
 #!/bin/python3
 
 from simpleline import App
-from simpleline.render import InputState
 from simpleline.render.prompt import Prompt
-from simpleline.render.screen import UIScreen
+from simpleline.render.screen import UIScreen, InputState
 from simpleline.render.widgets import TextWidget, CenterWidget
